TASK/Task_7/2_ecommerce_api/requirements.txt
fastapi
uvicorn
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart